data_pipeline/requirements.txt
# ──────────────────────────────────────────────────────────────
# Full requirements (pipeline + existing AI src/)
# ──────────────────────────────────────────────────────────────

# ── Pipeline ──────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.44.0
minio>=7.2.0
neo4j>=5.19.0
python-dotenv>=1.0.0

# ── AI / src ──────────────────────────────────────────────────
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
redis>=5.0.0
prometheus-client>=0.20.0
pydantic>=2.7.0

# ── Dev / test ────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
